% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{gos}
\alias{gos}
\title{Gos object}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
gos
}
\description{
Uses the reticulate framework to access the Gos API.
}
\details{
The Gos Python package is exposed through the \code{gos} object.
You can create and add to chart using its methods and classes,
as outlined in the
\href{https://gosling-lang.github.io/gos}{Gos Python documentation}.

In this package, use the \code{$} operator wherever you see the \code{.} operator
used in Python.
}
\seealso{
\href{https://gosling-lang.github.io/gos}{Gos Python documentation}
}
\keyword{datasets}
