% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{gosling_version}
\alias{gosling_version}
\title{Installed versions of Gos}
\usage{
gosling_version()
}
\value{
named \code{list} of version tags
}
\description{
Returns a named list of version tags for Gos
}
\examples{
if (interactive()) {
  gosling_version()
}

}
